refactor: :recycle: Moved popup creation to new module
